requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
asgiref==3.2.10           # via django
click==7.1.2              # via pip-tools
dj-rest-auth==1.1.2       # via -r requirements.in
django-cors-headers==3.5.0  # via -r requirements.in
django-filter==2.4.0      # via -r requirements.in
django-redis==4.12.1      # via -r requirements.in
django==3.1.2             # via -r requirements.in, dj-rest-auth, django-cors-headers, django-filter, django-redis, djangorestframework, djangorestframework-simplejwt
djangorestframework-link-header-pagination==0.1.1  # via -r requirements.in
djangorestframework-simplejwt==4.4.0  # via -r requirements.in
djangorestframework==3.12.1  # via -r requirements.in, dj-rest-auth, djangorestframework-link-header-pagination, djangorestframework-simplejwt
gunicorn==20.0.4          # via -r requirements.in
importlib-metadata==2.0.0  # via markdown
markdown==3.2.2           # via -r requirements.in
pip-tools==5.3.1          # via -r requirements.in
psycopg2-binary==2.8.6    # via -r requirements.in
pyjwt==1.7.1              # via djangorestframework-simplejwt
pytz==2020.1              # via django
pyyaml==5.3.1             # via -r requirements.in
redis==3.5.3              # via -r requirements.in, django-redis
six==1.15.0               # via pip-tools
sqlparse==0.3.1           # via django
zipp==3.2.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
